refactor(*): some refactoring
